requirements.txt
torch>=2.1.0
diffusers @ git+https://github.com/huggingface/diffusers.git@v0.21.4
transformers @ git+https://github.com/huggingface/transformers.git@v4.34.1
accelerate @ git+https://github.com/huggingface/accelerate.git@v0.23.0
compel>=2.0.2
numpy
torchsde
torchvision
ninja
safetensors
Pillow
waitress
timm
Flask
xformers
triton
bitsandbytes
omegaconf